Update README.md
just some markdown formatting to improve readability of the links.  Judging by the way you had formatted it originally I think you intended to have it as a list anyway?  Can cancel this PR if that wasn't the case.

Also used this as an exercise for following your steps outlined in "Contributing"