Add support for build modes  (#321)
Adding support for build modes. Initial support the following build modes:

Translated - Running a build with 'translationfile' as a compiler feature
Clean - Applying a set of preprocessorsymbols during compilation. The list of preprocessorsymbols is defined in the settings.json of the AL-GO project
Default - Running a build with the code as it is.
Default build mode is 🥁... Default