Add new apps ahead of .github and .AL-Go (#599)
Currently, app folders created using the AL Go actions will always be
appended at the end of the folder list in the workspace settings.
However, for most AL developers, access to the `.github` and `.AL-Go`
folders will be low priority. Additionally, some VS Code extensions
assume the relevant workspace folder to be the first one, which by
default is `.AL-Go`.

When adding a new app, this feature locates both `.github` and `.AL-Go`
in the workspace settings, determines which one appears earlier and
inserts the newly created app ahead of it. If neither can be found, it
will be appended at the end of the list, just like the current behavior.

With this, the first app created will automatically be at first
position, with subsequent apps following after.

---------

Co-authored-by: Freddy Kristiansen <freddy.kristiansen@microsoft.com>