Fix issue with getting dependencies in incremental build (#707)
Issue 1: after
https://github.com/microsoft/AL-Go/commit/7e245d8413b88c59b7d0a05bb4582ac107f99b84#diff-8b3bc43d627973d00981a6cb79e54160281d51f7a4a7aee41e856160e9841b24
there was an issue with `$_` not being defined.

Issue 2: `TestApp` mask is transformed to `<buildMode>TestApps` in
non-default mode. There are a few lines that needed fixing.

Note: The issue was not caught in E2E tests as there's no test to cover
the PR incremental build (where some projects are not built, but instead
fetched from a previous build).