Applying conditional settings for PR Builds (#958)
AL-GO doesn't apply conditional settings on PR Builds. In BCApps we have
the following conditional setting in the project settings

```
"buildModes": [ "Translated" ],
"ConditionalSettings": [
        {
            "branches": [
                "releases/*.[0-5]"
            ],
            "settings": {
                "buildModes": [ "Translated", "Strict"]
            }
        }
    ]
```

Yet when you see the PR builds, it doesn't build in StrictMode: 
https://github.com/microsoft/BCApps/actions/runs/8044365490

In ReadSettings we only apply conditional settings if the branchname
matches the pattern in the setting. The problem is that the branchName
is set to GITHUB_BASE_REF which for PR builds triggered by the
pull_request trigger is the merge branch (e.g. refs/heads/merge/123).
The branch pattern for the conditional setting will therefore never
match on PR Builds.