Add .pre-commit-config.yml file with pre-commit hooks (#955)
This pull request adds a `.pre-commit-config.yaml` file to the
repository, which includes pre-commit hooks for code formatting and
quality checks.

The hooks can be installed locally using [this
guide](https://pre-commit.com/#install) to check for and rectify any
violations before committing files, but using
[pre-commit.ci](https://pre-commit.ci/), it's also possible to register
this repository to be automatically corrected and updated by a bot. This
way, any pull requests for this project will either always adhere to
guidelines defined by the hooks, or fail a status check if they require
manual adjustments.

To enable _pre-commit.ci_, it's important to log into the service and
authorize access to this repository. Beyond that, no further action is
required.

Here's a list of all the hooks added with the `.pre-commit-config.yml`
file. All of them pass after a single autofix commit.

- executablebooks/mdformat/mdformat
- pre-commit/pre-commit-hooks/check-added-large-files
- pre-commit/pre-commit-hooks/check-case-conflict
- pre-commit/pre-commit-hooks/check-json
- pre-commit/pre-commit-hooks/check-xml
- pre-commit/pre-commit-hooks/check-yaml
- pre-commit/pre-commit-hooks/check-merge-conflict
- pre-commit/pre-commit-hooks/detect-private-key
- pre-commit/pre-commit-hooks/end-of-file-fixer
- pre-commit/pre-commit-hooks/trailing-whitespace
- pre-commit/pre-commit-hooks/mixed-line-ending
- pre-commit/pre-commit-hooks/sort-simple-yaml

Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>