Allow "ubuntu-latest" runners for all workflows (#1146)
This pull request adds a check to allow updating the `runs-on` and
`shell` parameters of all workflows. Previously, the workflows "Update
AL Go System Files" and "Pull Request Handler" were excluded when
changing the two parameters to prevent deadlocks and security issues.
With this change, a check will only allow changing the parameters of the
blacklisted workflows if the `runs-on` parameter is either
`windows-latest` or `ubuntu-latest`. Additionally, a check in the
function `ModifyRunsOnAndShell` ensures only the value `pwsh` is allowed
when using an `ubuntu-latest` runner.

---------

Co-authored-by: Freddy Kristiansen <freddy.kristiansen@microsoft.com>
Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>
Co-authored-by: freddydk <freddydk@users.noreply.github.com>