Install dependencies (#1225)
Fixes #1215 

Add new DeployTo setting called DependencyInstallMode, which can be
ignore, install or upgrade. Default is install (which doesn't cause
disruption with prior functionality)

If DependencyInstallMode is set to ignore - then dependencies will not
be installed/upgraded.

All AppSource Apps will be installed/upgraded first, which can be done
in any order as they are installed/upgraded with
`-installOrUpdateNeededDependencies`. All dependency PTEs are installed
afterwards in dependency order.

Rules for deploying dependencies to online environments goes here:

If the App (from dependencies) is a PTE:
- If it is installed in a newer version - leave it and see if it works
- If it is installed in an older version - Upgrade if
DependencyInstallMode is upgrade (else issue WARNING)

If the App (from dependencies) is an AppSource App:
- If it isn't already installed - install it from AppSource
- If it is installed in an older version in the Dev scope - issue a
warning and leave as is
- If it is installed in an older version - Upgrade if
DependencyInstallMode is upgrade (else issue WARNING)
- If it is installed in a newer version - issue a warning and leave as
is

TODO:
- [x] Add Release note
- [x] Add to settings.md

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>