Versioning Strategy 3 (#1356)
Setting versioning strategy to 3 will allow 3 segments of the version
number to be defined in app.json and repoVersion. Only the 4th segment
(Revision) will be defined by the GitHub
[run_number](https://go.microsoft.com/fwlink/?linkid=2217416&clcid=0x409)
for the CI/CD workflow. Increment version number and Create Release now
also supports the ability to set a third segment to the RepoVersion and
appversion in app.json.

Also, add support for skipping updating the dependency version numbers
in apps.

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>