Remove thisbuild artifacts and support preprocessor symbols (#1374)
This is a proposed fix for
https://github.com/orgs/microsoft/projects/521/views/1?pane=issue&itemId=91449371
and will greatly improve the capabilities of build modes (as requested
by partners during our roundtable at Directions EMEA).
 
### New Repository Settings

-
[`shortLivedArtifactsRetentionDays`](https://aka.ms/algosettings#shortLivedArtifactsRetentionDays)
determines the number of days to keep short lived build artifacts (f.ex
build artifacts from pull request builds, next minor or next major
builds). 1 is default. 0 means use GitHub default.
-
[`preProcessorSymbols`](https://aka.ms/algosettings#preProcessorSymbols)
is a list of preprocessor symbols to use when building the apps. This
setting can be specified in workflow specific settings files or in
conditional settings.

### Change in published artifacts

When using `useProjectDependencies` in a multi-project repository, AL-Go
for GitHub used to generate short lived build artifacts called
`thisBuild-<projectnaame>-<type>-...`. This is no longer the case.
Instead, normal build artifacts will be published and used by depending
projects. The retention period for the artifacts generated are
controlled by a settings called
[`shortLivedArtifactsRetentionDays`](https://aka.ms/algosettings#shortLivedArtifactsRetentionDays).

### Preprocessor symbols

It is now possible to define preprocessor symbols, which will be used
when building your apps using the
[`preProcessorSymbols`](https://aka.ms/algosettings#preProcessorSymbols)
setting. This setting can be specified in workflow specific settings
file or it can be used in conditional settings.

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>