Support authenticating as GitHub Apps instead of using Personal Access Tokens (#1311)
This PR is the first part of supporting GitHub Apps for authentication
instead of PATs in GhTokenWorkflow.

After this PR, we still need a PAT for running the End 2 End tests, and
we will have to have a PAT for End 2 End tests until we no longer
support the versions, which doesn't support using GitHub Apps for
authentication. The full scenario end 2 end test does use the GitHub app
for authentication and does as such test that GhTokenWorkflow can be a
GitHub app.

Also, the microsoft owned repositories used in end 2 end (bcsamples-*)
are also using the PAT, as we currently do not have the GitHub App
installed in the microsoft org. This will be part of a second part of
this - to use the PAT in fewer places in the end 2 end tests.

TODO:
- [x] Documentation

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>
Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>