Incremental Builds (#1304)
## Support incremental builds

### TODO

- [x] Add release notes
- [x] Add settings docs
- [x] Add end 2 end tests
- [x] Fix existing CI tests

### Use git diff instead of API

In v6.1 and earlier versions, we use the GitHub API to determine last
known good build and determine modified files in the Pull Request. It is
however possible to have done modifications to the main branch after the
last known good build - the changes will not be included in the PR and
this the PR builds will only include the modified files in the PR
together with the last known good build - potentially missing files.
This version locates the SHA, which was used to build the last known
good build and uses git diff to determine all changes from that build
including the changes in the PR to build what's needed.
git diff also supports many more files than the API, is faster and it
doesn't count as API calls against the GitHub limit.

### Changed <workflow>Schedule to workflowSchedule

Instead of using dynamic setting key name - use a workflow specific
setting and deprecate the old keys.

### Add concurrency protection to workflows

Add new workflow specific setting called - workflowConcurrency. It is
used to control concurrency of workflows.
Like with the `WorkflowSchedule` setting, this setting should be applied
in workflow specific settings files or conditional settings. By default,
all workflows allows for concurrency, except for the Create Release
workflow. If you are using incremental builds in CI/CD it is also
recommented to set WorkflowConcurrency to:

```
[
  "group: ${{ github.workflow }}-${{ github.ref }}",
  "cancel-in-progress: true"
]
```
in order to cancel prior incremental builds on the same branch.
Read more about workflow concurrency
[here](https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/control-the-concurrency-of-workflows-and-jobs).

### Remove warnings that tests doesn't exist if doNotBuildTests is true

No reason to warn about missing test apps if you don't want to build
them anyway

### Remove warnings and error when locating previous builds

Getting rid of warnings like this:

![{224ABFB5-16B3-496B-B3F7-F33F474185F0}](https://github.com/user-attachments/assets/4dfe3748-3114-4f5d-b193-0fe14d42375e)
due to DownloadProjectDependencies downloads artifacts double

and errors like:

![{1C5434BC-A199-444C-AAD7-CA1362EA7D61}](https://github.com/user-attachments/assets/47b72ffc-751d-4432-b607-671a45fdd01c)
if there wasn't a successful build in the repo.

### Check whether a package has been delivered to GitHubPackages or
NuGet before delivering

If the package already exists in the current version it has already been
delivered (might be from an earlier build)

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>
Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>