Run "Update AL-Go System Files" on multiple branches (#1037)
Issue: When scheduling "Update AL-Go System Files" workflow, it runs
only on the default branch (GitHub Actions limitation).

This PR alters the workflow to be able to schedule "Update AL-Go System
Files" to run on multiple branches.
Manually dispatching the workflow also allows selecting the branches to
run on (via a new input).

New property `workflowSchedule` setting, `includeBranches`, is added to
specify the branches when running the workflow on a schedule.

---------

Co-authored-by: Freddy Kristiansen <freddy.kristiansen@microsoft.com>
Co-authored-by: Alexander Holstrup <117829001+aholstrup1@users.noreply.github.com>