Downloaded apps no longer published in PR builds (#1481)
As part of the incremental builds feature, we download any apps that
have not been updated. In single project repos, these downloaded apps
are incorrectly published as artifacts in PR builds.

This PR adds a small snippet to remove any downloaded apps after the
build, to ensure only compiled apps are published.

---------

Co-authored-by: Freddy Kristiansen <freddy.kristiansen@microsoft.com>