Fix incremental build problems (#1488)
It can be hard to discover the reason why a full build was triggered
deep inside the log.
This PR adds a notice to the build when a full build is triggered, with
the reason behind it.

Fixes #1487

Additionally, a Pull Request, which should trigger a full build (due to
files changed) wouldn't trigger the change due to a parameter error.

Thirdly - in some cases, we would attempt to do a git diff with an empty
sha.

Lastly - skip the build cleanup step if no build step was ever run
(saves ~10 seconds)

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>