Legacy code signing (#1543)
Legacy code signing stopped working

Base64 version of the encrypted version of secrets could be visible in
the log. The secret would not be exposed as by using Base64 decoding you
would get the encrypted version of the secret, which only can be
decrypted on the runner on which it was encrypted - only if you have
access to run code on that machine, you would be able to decrypt the
secret.

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>