Page Scripting test result visualizer & updated test analyzer logic. (#1423)
I added functionality to visualize page scripting tests, similar to how
app and bcpt tests are visualized as MD.

I also updated the test analyzer action to take a new param, which
decides which of those 3 types of tests to analyze and visualize.
I think this is a better approach that a separate action for analyzing
page scripting tests, as it is nearly identical in implementation.

This approach also splits the analyzer visualization step into 3
separate steps, for each of the 3 types. This have the advantage, that
we can easily add new test types to the analyzer in the future, but we
are also more likely to display the results in the build summary. As
each step has a limit of 1MiB, the current approach of displaying both
app and bcpt results in the same step, are more likely to reach that
limit. Now each summary for each type have a separate limit of 1MiB.

---------

Co-authored-by: Freddy Kristiansen <freddy.kristiansen@microsoft.com>