Check release tag when creating a release (#1656)
This PR will check whether the latest release tag is lower than the
release tag specified.

If this is not the case, you will get an error or a warning depending on
whether you are creating a release based on the latest build (the
default).

Fixes #1627

Also includes fix for when no files modified on Git, deployment fails

Fixes #1657

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>