Dependencies to pre-published apps doesn't work in AL-Go for GitHub (#1661)
This pull request includes changes to improve the handling of missing
dependencies in the `RunPipeline.ps1` script

If an app has a dependency on an app, which is pre-published to the
container, you cannot specify that app in installApps, in fact, you
cannot install it at all.

This fix enables allows you to have dependencies to microsoft apps (from
the current version) without having to add these to installApps - just
add the dependency.

Fixes #1530

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>