check rate limit during invokewebrequest (#1663)
Use the GITHUB_TOKEN for read if the template repository is public and
the GhTokenWorkflow if the template repository is private/internal (with
the template repo as the only target). Rate limit for GitHub token
should be 1000 per repository per hour
https://docs.github.com/en/rest/using-the-rest-api/rate-limits-for-the-rest-api?apiVersion=2022-11-28#primary-rate-limit-for-github_token-in-github-actions

Beside this - the low level InvokeWebRequest will wait for rate limit
reset if rate limit is encountered and also handle 404 with redirect and
401 errors better, when trying to read a public repo with an access
token, which doesn't have access.

Fixes #1644 
Fixes #1478

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>