Replace E2EPAT with GitHub App  (#1717)
The purpose of this PR is to get rid of the E2EPAT used for running E2E
tests. It will instead be replaced by a GitHub app installed on the
BusinessCentralApps organization.

* **E2ETestHelper**: Introducing a new function refreshing the access
token and calling that function periodically when we need to use the
token.
* **Upgrade Tests**: Upgrade tests for AL-Go versions < 6.4 used to use
E2EPAT as its GHTokenWorkflow. Now they will use an access token
instead.
* **Federated Credentials Scenario**: This used to run on a repo in the
Microsoft org. Now we'll create a copy in the BusinessCentralApps
organization. This to avoid having to install the new GitHub app on the
Microsoft org. It needs quite elevated permissions so it'd preferable to
keep it in its own GitHub org.

Future work:
* **GitHub packages Scenario**: Here we will still need a PAT due to a
limitation on [GitHub's
side](https://docs.github.com/en/packages/learn-github-packages/about-permissions-for-github-packages#about-scopes-and-permissions-for-package-registries).
For now, this PR replaces the E2EPAT with a GitHubPackagesPAT which only
has read/write access to GitHub packages.
* **Access token scope** Currently the access tokens we request have all
permissions to all repositories the GitHub app allows. This can likely
be improved such that we request the minimum permissions needed for the
repositories we need to access. I would like to address this in a
separate PR though.


Related to
[AB#573954](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/573954)