Refactor CreateRelease and fix issue 1685 (#1686)
CreateRelease.yaml had a large portion of powershell in the .yaml file.
This has been moved out into an action and I have added an error if this
situation occurs:
https://github.blog/changelog/2023-11-02-github-actions-enforcing-workflow-scope-when-creating-a-release/
- so that people get a better error message than the current: HttpError:
Resource not accessible by integration - even I cannot remember the
reason for this when I see it again and again.

Fixes #1685

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>