Fixed access issue when Publishing from PR in private repo. (#1786)
When using the Pubish to Environment workflow, targeting a PR in a
private repo, the flow will fail with a permission issue, as the Github
token does not have access to read the PR and PR annotations.
I have updated the permissions of the workflow to allow the Github token
to access these.

Issue #1787