Add support for custom templates + customizations (#853)
Changes included in this PR:
- Support custom template repository = Point to an existing AL-Go
repository as a template, like:


![image](https://github.com/microsoft/AL-Go/assets/10775043/683a394b-3a9e-42b7-acce-8ab39b76a74f)

In this case, org/myptetemplate is a standard AL-Go repository, which
gets updated as any other AL-Go repository using Update AL-Go System
Files. This is a way of standardizing your repository based on type
specific AL-Go templates (which can be private or public).

Normally, Update AL-Go System Files follows this mechanism:

| File Type | Repository file(s) | AL-Go Template |
| :-- | :-- | :-: |
| Workflows | .github/workflows/*.yaml | <- |
| Repo Scripts | .github/*.ps1 | <- |
| Project Scripts | .AL-Go/*.ps1 | <- |

Workflows and scripts are read from the template repository and modified
based on settings and saved to the repository.

Using a custom template repository, the picture looks like this:

| File Type | Repository file(s) | Custom Template | AL-Go Template |
| :-- | :-- | :-: | :-: |
| Workflows | .github/workflows/*.yaml | <- Customizations | <- |
| Workflows | .github/workflows/*.yaml | <- New files | |
| Repo Scripts | .github/*.ps1 | | <- |
| Repo Scripts | .github/*.ps1 | <- New files | |
| Repo Settings | .github/templateRepoSettings.doNotEdit.json | <-
.github/AL-Go-Settings.json | |
| Project Scripts | .AL-Go/*.ps1 | | <- |
| Project Scripts | .AL-Go/*.ps1 | <- New files | |
| Project Settings | .github/templateProjectSettings.doNotEdit.json | <-
.AL-Go/settings.json | |

Workflows and scripts are read from the "real" template repository and
modified based on settings. Customizations for workflows in the custom
template repository are applied.

New workflows and scripts from the custom template repository are copied
over.

Repo and Project settings from the custom template repository are copied
over to new settings files, which are applied before the final repos
repo and project settings in order to allow the indirect template to add
new defaults to settings to be applied to all repositories.

The final repository can also contain customizations to workflows and
the winner here is the indirect template repository (if a custom job
exists in both places).

This PR: https://github.com/freddydk/customized/pull/17 - is an example
of a repository with customizations of CI/CD and _BuildALGoProject,
which runs Update AL-Go System Files and gets customizations from the
indirect template repository (both settings, scripts and workflow
customizations)

Documentation on customization capabilities is here:
https://github.com/freddydk/AL-Go/blob/customize/Scenarios/CustomizingALGoForGitHub.md

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>
Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>
Co-authored-by: Patrick Schiefer <35697862+PatrickSchiefer@users.noreply.github.com>
Co-authored-by: Philipp Stöhr <52424849+ps610@users.noreply.github.com>
Co-authored-by: Johannes Wikman <jwikman@users.noreply.github.com>
Co-authored-by: Alexander Holstrup <117829001+aholstrup1@users.noreply.github.com>