AL code analysis tracking in GitHub. (#1863)
Adding new feature to track AL code analysis results in GitHub.

We already have existing CodeCop features for AL in AL-Go, but this
builds on top of that and integrates any code scan issues with GitHub
security.

This allows alerts to appear in the Security tab of GitHub, and as
automated comments on PRs if new alerts are detected.

The SARIF upload is placed in a new job instead of the build job as only
one SARIF upload is allowed per sha. If multiple builds run in parallel,
like in a repo with several projects, then the SARIF uploads would
overwrite each other. The solution is to have a new job that downloads
all ErrorLogs from artifacts from all builds and combines them in a
single upload.

---------

Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>
Co-authored-by: Alexander Holstrup <117829001+aholstrup1@users.noreply.github.com>