Use Runner_Temp folder whenever available (#2017)
### ❔What, Why & How

<!-- Include description of the changes that will help reviewers in
their task -->
This PR changes the behavior in AL-Go to use the RUNNER_TEMP folder
whenever available. RUNNER_TEMP is a temporary directory on the runner.
This directory is emptied at the beginning and end of each job. That way
we ensure files added by AL-Go to the runner is automatically cleaned up
again.

Additionally, when creating temp folders, the folders will have shorter
paths. We've seen issues in the past because temp files/folders had too
long paths e.g.
https://github.com/microsoft/BCApps/actions/runs/19102211464


Related to issue:
[AB#612647](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/612647)

### ✅ Checklist

- [ ] Add tests (E2E, unit tests)
- [x] Update RELEASENOTES.md
- [ ] Update documentation (e.g. for new settings or scenarios)
- [ ] Add telemetry

<!-- Include more checklist entries, if needed -->

---------

Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>