Add telemetry for test results (#2048)
### ❔What, Why & How

<!-- Include description of the changes that will help reviewers in
their task -->
It might be interesting to track the number of tests being run in your
workflow over time, and to see whether tests are failing.

This PR would add that to the existing AL-Go telemetry. In the near
future we could look into adding a page to the starter-dashboard as
well, so an AL-Go partner could quite easily get an overview of their
test runs across all their repos.

### ✅ Checklist

- [ ] Add tests (E2E, unit tests)
- [x] Update RELEASENOTES.md
- [x] Update documentation (e.g. for new settings or scenarios)
- [x] Add telemetry

<!-- Include more checklist entries, if needed -->